requirements.txt
matplotlib
fastapi
sqlalchemy
requests
paho.mqtt
psycopg2
pydantic
uvicorn